refactor(e2ei): centralize and make crate-private mutating operations
We care about isolating those operations which mutate the database,
because they require a transaction.
