+++ e2e-identity/src/pki_env/crl.rs
@@ -1,12 +1,7 @@
 use std::collections::HashMap;
 
-use core_crypto_keystore::entities::E2eiCrl;
-
 use super::{Error, Result};
-use crate::{
-    pki_env::{PkiEnvironment, hooks::HttpMethod},
-    x509_check::revocation::PkiEnvironment as RjtPkiEnvironment,
-};
+use crate::pki_env::{PkiEnvironment, hooks::HttpMethod};
 
 impl PkiEnvironment {
     /// Fetch certificate revocation lists from the given URIs, return a map from the URLs to a DER-encoded certificate
@@ -31,16 +26,5 @@ impl PkiEnvironment {
         }
 
         Ok(crls)
-    }
-
-    /// Validate the CRL (trust anchors must be configured prior to this) and
-    /// save it to the database.
-    pub async fn save_crl(&self, crl_dp: &str, crl_der: &[u8]) -> Result<()> {
-        let crl = self.rjt_pki_env.lock().await.validate_crl_with_raw(crl_der)?;
-        let crl_data = E2eiCrl {
-            content: RjtPkiEnvironment::encode_crl_to_der(&crl)?,
-            distribution_point: crl_dp.to_owned(),
-        };
-        self.database.save(crl_data).await.map_err(Into::into)
     }
 }
